Revert "HOCS-2605 Cannot remove last nominated contact"
This reverts commit 4f88d07a6c1994fcee661ef82910f58fd3935555.

+++ src/main/java/uk/gov/digital/ho/hocs/info/api/NominatedContactService.java
@@ -53,8 +53,14 @@ public class NominatedContactService {
     public void deleteNominatedContact(UUID teamUUID, UUID nominatedContactUUID){
         log.debug("Deleting nominated contact with uuid: {} from team {}", nominatedContactUUID, teamUUID);
 
-        NominatedContact nominatedContact = nominatedContactRepository.findByUuid(nominatedContactUUID);
-        nominatedContactRepository.delete(nominatedContact);
-        log.info("Deleted nominated contact uuid: {} ", nominatedContactUUID);
+        Set<NominatedContact> contacts = nominatedContactRepository.findAllByTeamUUID(teamUUID);
+        if (contacts.size() > 1 ) {
+            NominatedContact nominatedContact = nominatedContactRepository.findByUuid(nominatedContactUUID);
+            nominatedContactRepository.delete(nominatedContact);
+            log.info("Deleted nominated contact uuid: {} ", nominatedContactUUID);
+        } else {
+            throw new ApplicationExceptions.NominatedContactDeleteException(
+                    "Unable to delete nominated contact for team {} - teams must have at least one contact", teamUUID);
+        }
     }
 }
